chore(samples): regenerate cli

+++ samples/cli/src/main/java/net/accelbyte/sdk/cli/api/eventlog/event/GetEventByEventTypeAndEventIDHandler.java
@@ -44,13 +44,13 @@ public class GetEventByEventTypeAndEventIDHandler implements Callable<Integer> {
     String namespace;
 
     @Option(names = {"--offset"}, description = "offset")
-    Float offset;
+    Integer offset;
 
     @Option(names = {"--endDate"}, description = "endDate")
     String endDate;
 
     @Option(names = {"--pageSize"}, description = "pageSize")
-    Float pageSize;
+    Integer pageSize;
 
     @Option(names = {"--startDate"}, description = "startDate")
     String startDate;
