+++ samples/cli/src/main/java/net/accelbyte/sdk/cli/api/eventlog/event_v2/QueryEventStreamHandler.java
@@ -41,10 +41,10 @@ public class QueryEventStreamHandler implements Callable<Integer> {
     String endDate;
 
     @Option(names = {"--offset"}, description = "offset")
-    Float offset;
+    Integer offset;
 
     @Option(names = {"--pageSize"}, description = "pageSize")
-    Float pageSize;
+    Integer pageSize;
 
     @Option(names = {"--startDate"}, description = "startDate")
     String startDate;
