+++ samples/cli/src/main/java/net/accelbyte/sdk/cli/api/eventlog/user_information/GetUserActivitiesHandler.java
@@ -41,10 +41,10 @@ public class GetUserActivitiesHandler implements Callable<Integer> {
     String userId;
 
     @Option(names = {"--offset"}, description = "offset")
-    Float offset;
+    Integer offset;
 
     @Option(names = {"--pageSize"}, description = "pageSize")
-    Float pageSize;
+    Integer pageSize;
 
 
     @Option(names = {"--logging"}, description = "logger")
